% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{Overall App Server}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}
}
\value{
shiny server objects
}
\description{
Overall App Server
}
