% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleServer.R
\name{moduleServer}
\alias{moduleServer}
\title{Module Server wrapper}
\usage{
moduleServer(id, module, session = getDefaultReactiveDomain())
}
\arguments{
\item{id}{}

\item{module}{}

\item{session}{}
}
\value{

}
\description{
Starting in Shiny 1.5 the recomdation will be to use a method for definding the server that calls a functionand will include a new function moduleServer that is below instead of use callModule that was in Shiny 1.4  
Reference: https://github.com/rstudio/shiny/blob/master/R/modules.R
For an example see https://mastering-shiny.org/action-modules.html
}
