% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptionsUI.R
\name{OptionsServer}
\alias{OptionsServer}
\title{Server for options module}
\usage{
OptionsServer()
}
\value{
moduleServer for options module - currently and empty shell
}
\description{
Server for options module
}
