% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_simulation.R
\name{SimulationIntroUI}
\alias{SimulationIntroUI}
\title{Simulation shiny module UI - Intro subtab}
\usage{
SimulationIntroUI(strID)
}
\arguments{
\item{strID}{namespace id}
}
\value{
UI for intro subtab of simulation module
}
\description{
Simulation shiny module UI - Intro subtab
}
