% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_simulation.R
\name{SimulationProgramServer}
\alias{SimulationProgramServer}
\title{Simulation Shiny Module Server}
\usage{
SimulationProgramServer()
}
\value{
Shiny module
}
\description{
Simulation Shiny Module Server
}
