% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_simulation.R
\name{SimulationUI}
\alias{SimulationUI}
\title{Simulation shiny module UI}
\usage{
SimulationUI()
}
\value{
shiny module
}
\description{
For complex tabs it may be best to use an approach similar to this tab.   Have functions for each subtab of the tab that each reside in a seperate file. Then create a functionUI that would be the tab and all subtabs, in this example SimulationUI.   This approach keeps the main UI from having to call all subtab module UI functions. Note: for this simple tab  all UI functions for this tab are in this one file but in practice it may be best to seperate each subtab funciton to different files.
}
