% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_themeSwitcher.R
\name{ThemeSwithcherUI}
\alias{ThemeSwithcherUI}
\title{Theme switcher module UI}
\usage{
ThemeSwithcherUI(dropDownLabel = "Change Theme", defaultTheme = "grey_light")
}
\arguments{
\item{dropDownLabel}{}

\item{defaultTheme}{}
}
\value{
UI for theme selector
}
\description{
Theme switcher module UI
}
